Input file list for realignmenttargetcreator working
The flatten() op is not working for my channel in the
way I expected, so I needed to workaround that to grab and merge
everything by sample/subject Id (grabbing element [0] from
duplicated list). Also using a List collect() and join() to
make the GATK command line input work.

Now the realignment target creation and realignment steps
"work", but the output is not properly grabbed by NF,
need to split the *.real.bam files from list in the
realignment output channel.