A directory of FASTQs can be given with --sampleDir
This adds the option `--sampleDir`, which is mutually exclusive with
`--sample` and `--test` (only one of the three may be given).

Wildcards are supported, so `'--sampleDir=my/directory/with/samples/*'` works.
Each directory is then treated as a single patient with a single sample.

Also, add a function `flowcellIdFromFastq()1, which extracts the flowcell ID
from a FASTQ file. This is at the moment only used with `--sampleDir`, not
when on `--sample` or `--test` is provided.