Simplify defineReferenceMap
The ternary with the nested Elvis operator can be factored out.
A simple if makes it clearer what is going on.

Note that map[key] (or map.key) returns null if the key does not exist.
null has the same truth value as false, so no need to convert it to false.