Remove -XL parameters to GATK tools
`-XL` is `--excludeIntervals` and it was set to `-XL hs37d5 -XL NC_007605` in
order to exclude decoys and the Epstein-Barr sequence. These names are specific
to GRCh37, however. With this change, we use `-L` instead, we gives a list of
all intervals to *include*`. For some tools that already before got an explicit
list of intervals to work on, this change just removes the `-XL` parameters as
they should not have had any effect anyway.