Add a RunGenotypeGVCFs process
The process gets HaplotypeCaller’s genomic VCFs (.g.vcf) files as
input and converts them to regular VCFs. This is done on the interval-split
VCF files. Both the normal and the genomic VCF are then passed to ConcatVCF
in order to create whole-genome VCFs. The output is written to
VariantCalling/HaplotypeCaller.

Closes #275