Avoid creating huge recalibrated BAM files
When HaplotypeCaller is the only requested tool, the pipeline now skips the
RecalibrateBam process.

The option --BQSR is directly passed to HaplotypeCaller. This means that
HaplotypeCaller does the base quality recalibration "on the fly". The results
should be the same as before.

To clarify: Base quality recalibration is still done, we just never create
the BAM file with recalibrated bases.

PrintReads, which was previously used to create the recalibrated BAM file,
took around 19 hours. With this change, and when only HaplotypeCaller is
run, these 19 hours are saved. HaplotypeCaller is a bit slower because it
needs to recalibrate bases on-the-fly , but this is distributed over all
HaplotypeCaller processes running in parallel (working on separate
intervals). For calling variants on a full test genome, HaplotypeCaller now
needs 12 instead of 11 hours. In total, we therefore save 18 hours.

Things to note:

- The base-recalibrated BAM file is still created if HaplotypeCaller is not
  the only requested tool.
- --BQSR is *always* passed to HaplotypeCaller, even when a recalibrated BAM
  is created.
- The same --BQSR option is (likely) also supported by MuTect and MuTect 2,
  so the same optimization could be applied.

Closes #281