Concatenate VCFs from different intervals much faster
Instead of using Picard’s SortVcf, we can simply concatenate the VCFs in the
correct order. This works because we know that the VCFs do not overlap each
other and because each of them is already sorted.

The ConcatVcf process previously took around 4-5 hours. One hour was spent
in concatenating all files (with awk), much of the rest on Picard’s SortVcf,
and some time on compressing the output.

With this change, ConcatVcf takes less than 5 minutes even on a large
collection of GVCFs.