Make --step=skippreprocessing work
This did not play well with skipping BQSR and nothing was actually run.

Now, if --step=skippreprocessing is provided, we assume that the input
is a recalibrated BAM. HaplotypeCaller does not get the --BQSR option
in this case and therefore will not attempt to re-recalibrate).

Use --step=recalibrate to start from a non-recalibrated BAM.

Fixes #351