Sort interval chunks by estimated runtime
Previously, we relied on an appropriately sorted input file for this.
Now we get the estimated duration from the input BED and sort it
ourselves. If there is no runtime estimate for a record in the BED,
the fallback is to assume a speed of 1000 nucleotides per second.
If no runtimes are annotated, then the interval groups are effectively
grouped by length.