ci: Set fail-fast to false
This will keep tests from all failing when one fails. It'll waste a
bunch of CI hours when things are more stable, but it will make sure
every test in the matrix gets ran.

This might be a good setting to keep on though, if all the tests fail
out, it's hard to pick out which one passed. I think we might need to
combo these with the paths-filter

Co-authored-by: FriederikeHanssen <Friederike.hanssen@qbic.uni-tuebingen.de>