Swap sentieon to nf-test and fix (#1927)
Swap sentieon to nf-test and fix.
The modules update changed the channels on a few of the modules, this
wasn't reflected in the code.
I had to change the options in `SENTIEON_VARCAL_SNP` to get that to work
in the test without it complaining, I set that in the
tests/nextflow.config file.

- [ ] This comment contains a description of changes (with reason).
- [ ] If you've fixed a bug or added code that should be tested, add
tests!
- [ ] If you've added a new tool - have you followed the pipeline
conventions in the [contribution
docs](https://github.com/nf-core/sarek/tree/master/.github/CONTRIBUTING.md)
- [ ] If necessary, also make a PR on the nf-core/sarek _branch_ on the
[nf-core/test-datasets](https://github.com/nf-core/test-datasets)
repository.
- [ ] Make sure your code lints (`nf-core pipelines lint`).
- [ ] Ensure the test suite passes (`nextflow run . -profile test,docker
--outdir <OUTDIR>`).
- [ ] Check for unexpected warnings in debug mode (`nextflow run .
-profile debug,test,docker --outdir <OUTDIR>`).
- [ ] Usage Documentation in `docs/usage.md` is updated.
- [ ] Output Documentation in `docs/output.md` is updated.
- [ ] `CHANGELOG.md` is updated.
- [ ] `README.md` is updated (including new tool citations and
authors/contributors).

---------

Co-authored-by: Maxime U Garcia <max.u.garcia@gmail.com>
Co-authored-by: nf-core-bot <core@nf-co.re>