Changes to UMI processing (#1937)
Big set of changes to UMI processing.
Currently the only UMI processing is through generating consensus reads
using fgbio, which is inefficient for whole genome sequencing.
Often the UMI might have already been extracted into the read header
(such as through bclconvert) or maybe in the bam file already.

Introduces:

- fastp for UMI extraction from reads (R1/R2/both or index reads in
fastq header) [test added]
- movement of UMIs from read name to RX: tag on bam/cram files, allowing
MarkDuplicates and Sentieon to get at them. This is required if
extracted using fastp, or if they were provided originally in that way.
[test added for MarkDuplicates]
- If using fgbio consensus, then this will now support UMIs in read
names.
- sentieon consensus mode (technically doesn't require UMIs, but while I
was fiddling with the config...)
- Ensures that fgbio merges together the lanes prior to consensus
generation (fixes #802). I also swapped to `samtools` rather than
`samblaster` to do the merge and sort, as is done in the fastquorum
pipeline; this removes samblaster entirely (I couldn't get samblaster to
merge the multiple files in one go).
- Uses Fulcrum Genomics plugin to validate the read structures, see
https://github.com/nf-core/fastquorum/pull/123
- If using sentieon dedup from bam/cram files without first aligning,
the UMIs will only be used if one of the two parameters above are set or
`umi_tag` parameter is set.
- MarkDuplicates will automatically use the RX tag if it is present in
the bam files, but sentieon does not.